+++ frontend/src/__tests__/Login.test.tsx
@@ -99,7 +99,7 @@ describe("Login page", () => {
         can_create_users: true,
         can_create_user_groups: true,
         can_create_connections: true,
-        can_create_connection_folders: true,
+        can_use_quick_share: true,
         can_create_sharing_profiles: true,
         can_view_sessions: true,
       },