+++ frontend/src/__tests__/api.test.ts
@@ -432,7 +432,7 @@ describe("API endpoint functions", () => {
       can_create_users: false,
       can_create_user_groups: false,
       can_create_connections: false,
-      can_create_connection_folders: false,
+      can_use_quick_share: false,
       can_create_sharing_profiles: false,
       can_view_sessions: false,
     });
@@ -448,7 +448,7 @@ describe("API endpoint functions", () => {
       can_create_users: false,
       can_create_user_groups: false,
       can_create_connections: false,
-      can_create_connection_folders: false,
+      can_use_quick_share: false,
       can_create_sharing_profiles: false,
       can_view_sessions: false,
     });